Added Vogel equation
In this commit I added the Vogel equation considering the bubble point pressure. So, when the bottomhole pressure is greater than the bubble pressure the code will use the pseudo-steady state Darcy equation. Otherwise, it will use the Vogel equation.
Finally, I did some corrections on the comments and created  other code lines.